Back Terminado

+++ src/main/java/api_reclamos_spring/service/IReclamoService.java
@@ -6,7 +6,7 @@ import api_reclamos_spring.model.entity.Reclamo;
 
 public interface IReclamoService {
 	public List<Reclamo> findAll();
-	public Reclamo findById(int id);
+	public Reclamo findById(Long id);
 	public void save(Reclamo reclamo);
 	public void update(int reclamoId, Reclamo reclamo);
 	public void deleteById(int id);
